safari-534.53-branch/LayoutTests/inspector/debugger/xhr-breakpoints-expected.txt

Tests XHR breakpoints.

Debugger was enabled.

Running: testXHRBreakpoint
Script execution paused.
Call stack:
    0) sendRequest (xhr-breakpoints.html:11)
    1)  (:1)
Script execution resumed.

Running: testPauseOnAnyXHR
Script execution paused.
Script execution resumed.
Script execution paused.
Script execution resumed.

Running: testDisableBreakpoint
Script execution paused.
Script execution resumed.
Debugger was disabled.
